+++ webui/src/admin/model.ts
@@ -1065,3 +1065,41 @@ export function statusTone(status: string): string {
       return "idle";
   }
 }
+
+export function getAvatarDisplayUrl(avatarUrl: string): string {
+  if (!avatarUrl) {
+    return "";
+  }
+  if (avatarUrl.startsWith("avatar://")) {
+    const avatarId = avatarUrl.substring(9);
+    return `/api/system/services/avatar/${avatarId}`;
+  }
+  return avatarUrl;
+}
+
+export function agentAvatarUrl(
+  agent: ServiceWithRuntime | null | undefined,
+): string {
+  if (!agent) {
+    return "";
+  }
+
+  if (agent.agent_type.type === "qq_chat") {
+    const profile = agent.qq_chat_profile;
+    const explicit = String(profile?.bot_avatar_url ?? "").trim();
+    if (explicit) {
+      return explicit;
+    }
+    const botUserId = String(profile?.bot_user_id ?? "").trim();
+    if (!botUserId) {
+      return "";
+    }
+    return `https://q1.qlogo.cn/g?b=qq&nk=${encodeURIComponent(botUserId)}&s=640`;
+  }
+
+  return getAvatarDisplayUrl(String(agent.avatar_url ?? ""));
+}
+
+export function agentInitial(name: string): string {
+  return (name || "B").trim().slice(0, 1).toUpperCase();
+}
