+++ webui/src/admin/view/Chat.vue
@@ -1,0 +1,2799 @@
+<template>
+  <section :class="embedded ? 'chat-embedded-wrapper' : 'page chat-page'">
+    <div v-if="!embedded" class="page-hero">
+      <h2>Chat</h2>
+    </div>
+    <div :class="embedded ? 'chat-embedded-inner' : 'chat-page-panel'">
+      <section class="panel chat-panel">
+        <div class="chat-toolbar">
+          <div class="chat-agent-picker">
+            <div class="chat-agent-picker-title">选择 Service</div>
+            <div class="chat-agent-cards">
+              <div
+                v-if="servicesLoading && services.length === 0"
+                class="chat-service-loading"
+                aria-live="polite"
+              >
+                <span class="chat-service-loading-spinner"></span>
+                <span>Service 加载中...</span>
+              </div>
+              <template v-else>
+                <button
+                  v-for="agent in services.filter((a) => chatEligibleServiceTypes.has(a.agent_type.type))"
+                  :key="agent.config_id"
+                  class="chat-agent-card"
+                  :class="{
+                    active: selectedServiceId === agent.config_id,
+                    inactive: agent.runtime.status !== 'running',
+                  }"
+                  @click="selectedServiceId = agent.config_id"
+                >
+                  <img
+                    v-if="agentAvatarUrl(agent)"
+                    class="chat-agent-card-avatar"
+                    :src="agentAvatarUrl(agent)"
+                    alt="agent avatar"
+                  />
+                  <div v-else class="chat-agent-card-avatar chat-agent-card-avatar--fallback">
+                    {{ agentInitial(agent.name) }}
+                  </div>
+                  <div class="chat-agent-card-meta">
+                    <strong>{{ agent.name }}</strong>
+                    <span>{{ readableAgentType(agent.agent_type.type) }}</span>
+                  </div>
+                  <span v-if="agent.runtime.status !== 'running'" class="agent-status-badge">
+                    未运行
+                  </span>
+                </button>
+              </template>
+            </div>
+          </div>
+          <button class="btn ghost" @click="reloadSessions">刷新历史</button>
+          <button
+            v-if="isWorkspaceService"
+            class="btn ghost"
+            :disabled="pickingDirectory"
+            @click="pickDirectory"
+          >
+            {{ pickingDirectory ? "选择中..." : "打开目录" }}
+          </button>
+        </div>
+
+        <div class="chat-layout">
+          <aside class="chat-sessions">
+            <div class="chat-sessions-header">历史</div>
+            <template v-for="group in groupedSessions" :key="group.pathKey">
+              <div class="chat-session-group-header" :title="group.path ?? undefined">
+                📁 {{ group.label }}
+              </div>
+              <div
+                v-for="session in group.sessions"
+                :key="session.session_id"
+                class="chat-session-item"
+                :class="{ active: session.session_id === activeSessionId }"
+              >
+                <button class="chat-session-main" @click="openSession(session.session_id)">
+                  <strong>{{ session.session_id.slice(0, 8) }}</strong>
+                  <span class="muted">{{ formatTime(session.updated_at) }}</span>
+                </button>
+                <button
+                  class="chat-session-delete"
+                  title="删除会话"
+                  @click.stop="removeSession(session.session_id)"
+                >
+                  ×
+                </button>
+              </div>
+            </template>
+            <div v-if="sessions.length === 0" class="muted">暂无历史会话</div>
+          </aside>
+
+          <div class="chat-main">
+            <div v-if="isWorkspaceService" class="workspace-path-display">
+              <span class="path-label">当前工作目录：</span>
+              <span class="path-value" :class="{ 'path-unset': !workspacePath }">
+                {{ workspacePath || '未选择工作目录' }}
+              </span>
+            </div>
+            <div class="chat-messages" ref="messagesContainer">
+              <div v-if="messages.length === 0" class="empty-state"></div>
+              <div
+                v-for="group in messageGroups"
+                :key="group.id"
+                class="chat-bubble-row"
+                :class="group.role"
+              >
+                <img
+                  v-if="group.role === 'assistant' && group.avatarUrl"
+                  class="chat-message-avatar"
+                  :src="group.avatarUrl"
+                  alt="bot avatar"
+                />
+                <div
+                  v-else-if="group.role === 'assistant'"
+                  class="chat-message-avatar chat-message-avatar--fallback"
+                >
+                  {{ agentInitial(group.agentName || "Bot") }}
+                </div>
+                <div v-if="group.role === 'assistant'" class="chat-bubble-col">
+                  <div
+                    v-for="(message, idx) in group.messages"
+                    :key="message.id + '-' + idx"
+                    class="chat-message-item"
+                  >
+                    <div
+                      v-if="
+                        idx === group.messages.length - 1 &&
+                        ((message.liveToolCalls && message.liveToolCalls.length > 0) ||
+                          message.toolCalls.length > 0 ||
+                          activeToolDetail?.messageId === message.id)
+                      "
+                      class="chat-tool-above-content"
+                    >
+                      <div
+                        v-if="message.liveToolCalls && message.liveToolCalls.length > 0"
+                        class="chat-tool-inline-list"
+                      >
+                        <div
+                          v-for="liveCall in message.liveToolCalls"
+                          :key="liveCall.call_id"
+                          class="chat-live-tool-wrapper"
+                        >
+                          <template
+                            v-if="
+                              classifyToolCall(liveCall.name, liveCall.arguments, liveCall.result).type ===
+                              'generic'
+                            "
+                          >
+                            <button
+                              class="chat-tool-inline"
+                              :class="{ active: expandedLiveToolCalls.has(liveCall.call_id) }"
+                              @click="toggleLiveToolCall(liveCall.call_id)"
+                            >
+                              <span v-if="!liveCall.done" class="live-tool-spinner"></span>
+                              <span v-else class="live-tool-done-icon">✓</span>
+                              工具调用: {{ liveCall.name }}
+                            </button>
+                            <div
+                              v-if="expandedLiveToolCalls.has(liveCall.call_id)"
+                              class="chat-tool-detail-inline"
+                            >
+                              <div class="chat-tool-detail-inline-block">
+                                <div class="chat-tool-detail-caption">arguments</div>
+                                <pre>{{ formatToolPayload(liveCall.arguments) }}</pre>
+                              </div>
+                              <div v-if="liveCall.done" class="chat-tool-detail-inline-block">
+                                <div class="chat-tool-detail-caption">result</div>
+                                <pre>{{ liveCall.result || "(空结果)" }}</pre>
+                              </div>
+                              <div v-else class="chat-tool-detail-inline-block">
+                                <div class="chat-tool-detail-caption">result</div>
+                                <div class="live-tool-pending">推理中...</div>
+                              </div>
+                            </div>
+                          </template>
+                          <ToolCallBadge
+                            v-else
+                            :kind="classifyToolCall(liveCall.name, liveCall.arguments, liveCall.result)"
+                            :loading="!liveCall.done"
+                            @click="
+                              liveCall.done &&
+                              openToolPreview(classifyToolCall(liveCall.name, liveCall.arguments, liveCall.result))
+                            "
+                          />
+                        </div>
+                      </div>
+                      <div v-if="message.toolCalls.length > 0" class="chat-tool-inline-list">
+                        <template v-for="toolCall in message.toolCalls" :key="toolCall.id">
+                          <button
+                            v-if="
+                              classifyToolCall(
+                                toolCall.function.name,
+                                toolCall.function.arguments,
+                                getToolResultText(toolCall.id),
+                              ).type === 'generic'
+                            "
+                            class="chat-tool-inline"
+                            :class="{ active: activeToolCallId === toolCall.id }"
+                            @click="openToolDetail(message.id, toolCall.id)"
+                          >
+                            调用工具: {{ toolCall.function.name }}
+                          </button>
+                          <ToolCallBadge
+                            v-else
+                            :kind="
+                              classifyToolCall(
+                                toolCall.function.name,
+                                toolCall.function.arguments,
+                                getToolResultText(toolCall.id),
+                              )
+                            "
+                            @click="
+                              openToolPreview(
+                                classifyToolCall(
+                                  toolCall.function.name,
+                                  toolCall.function.arguments,
+                                  getToolResultText(toolCall.id),
+                                ),
+                              )
+                            "
+                          />
+                        </template>
+                      </div>
+                      <div
+                        v-if="activeToolDetail?.messageId === message.id"
+                        class="chat-tool-detail-inline"
+                      >
+                        <div class="chat-tool-detail-inline-header">
+                          <strong>{{ activeToolDetail.toolCall.function.name }}</strong>
+                          <button class="chat-tool-detail-inline-close" @click="closeToolDetail">
+                            收起
+                          </button>
+                        </div>
+                        <div class="chat-tool-detail-inline-block">
+                          <div class="chat-tool-detail-caption">tool_call_id</div>
+                          <code>{{ activeToolDetail.toolCall.id }}</code>
+                        </div>
+                        <div class="chat-tool-detail-inline-block">
+                          <div class="chat-tool-detail-caption">arguments</div>
+                          <pre>{{ formatToolPayload(activeToolDetail.toolCall.function.arguments) }}</pre>
+                        </div>
+                        <div class="chat-tool-detail-inline-block">
+                          <div class="chat-tool-detail-caption">result</div>
+                          <pre>{{ activeToolDetail.result || "(空结果)" }}</pre>
+                        </div>
+                      </div>
+                    </div>
+                    <div
+                      v-if="message.thinkingContent"
+                      class="chat-thinking-block"
+                      :class="{ collapsed: !message.thinkingExpanded }"
+                    >
+                      <button
+                        class="chat-thinking-toggle"
+                        @click="message.thinkingExpanded = !message.thinkingExpanded"
+                      >
+                        <span class="chat-thinking-icon">
+                          {{ message.thinkingExpanded ? '▼' : '▶' }}
+                        </span>
+                        思考过程
+                        <span
+                          v-if="message.streaming && message.thinkingExpanded"
+                          class="live-tool-spinner"
+                        ></span>
+                      </button>
+                      <div v-if="message.thinkingExpanded" class="chat-thinking-content">
+                        {{ message.thinkingContent }}
+                      </div>
+                    </div>
+                    <div
+                      v-if="message.content.trim().length > 0 || message.streaming"
+                      class="chat-bubble"
+                      :class="message.role"
+                    >
+                      <div
+                        class="chat-bubble-content markdown-body"
+                        v-html="renderMessageContent(message.content, message.streaming)"
+                      ></div>
+                      <div class="chat-bubble-time">{{ formatChatTime(message.timestamp) }}</div>
+                    </div>
+                    <div
+                      v-if="
+                        idx !== group.messages.length - 1 &&
+                        ((message.liveToolCalls && message.liveToolCalls.length > 0) ||
+                          message.toolCalls.length > 0 ||
+                          activeToolDetail?.messageId === message.id)
+                      "
+                      class="chat-tool-below-content"
+                    >
+                      <div
+                        v-if="message.liveToolCalls && message.liveToolCalls.length > 0"
+                        class="chat-tool-inline-list"
+                      >
+                        <div
+                          v-for="liveCall in message.liveToolCalls"
+                          :key="liveCall.call_id"
+                          class="chat-live-tool-wrapper"
+                        >
+                          <template
+                            v-if="
+                              classifyToolCall(liveCall.name, liveCall.arguments, liveCall.result).type ===
+                              'generic'
+                            "
+                          >
+                            <button
+                              class="chat-tool-inline"
+                              :class="{ active: expandedLiveToolCalls.has(liveCall.call_id) }"
+                              @click="toggleLiveToolCall(liveCall.call_id)"
+                            >
+                              <span v-if="!liveCall.done" class="live-tool-spinner"></span>
+                              <span v-else class="live-tool-done-icon">✓</span>
+                              工具调用: {{ liveCall.name }}
+                            </button>
+                            <div
+                              v-if="expandedLiveToolCalls.has(liveCall.call_id)"
+                              class="chat-tool-detail-inline"
+                            >
+                              <div class="chat-tool-detail-inline-block">
+                                <div class="chat-tool-detail-caption">arguments</div>
+                                <pre>{{ formatToolPayload(liveCall.arguments) }}</pre>
+                              </div>
+                              <div v-if="liveCall.done" class="chat-tool-detail-inline-block">
+                                <div class="chat-tool-detail-caption">result</div>
+                                <pre>{{ liveCall.result || "(空结果)" }}</pre>
+                              </div>
+                              <div v-else class="chat-tool-detail-inline-block">
+                                <div class="chat-tool-detail-caption">result</div>
+                                <div class="live-tool-pending">推理中...</div>
+                              </div>
+                            </div>
+                          </template>
+                          <ToolCallBadge
+                            v-else
+                            :kind="classifyToolCall(liveCall.name, liveCall.arguments, liveCall.result)"
+                            :loading="!liveCall.done"
+                            @click="
+                              liveCall.done &&
+                              openToolPreview(classifyToolCall(liveCall.name, liveCall.arguments, liveCall.result))
+                            "
+                          />
+                        </div>
+                      </div>
+                      <div v-if="message.toolCalls.length > 0" class="chat-tool-inline-list">
+                        <template v-for="toolCall in message.toolCalls" :key="toolCall.id">
+                          <button
+                            v-if="
+                              classifyToolCall(
+                                toolCall.function.name,
+                                toolCall.function.arguments,
+                                getToolResultText(toolCall.id),
+                              ).type === 'generic'
+                            "
+                            class="chat-tool-inline"
+                            :class="{ active: activeToolCallId === toolCall.id }"
+                            @click="openToolDetail(message.id, toolCall.id)"
+                          >
+                            调用工具: {{ toolCall.function.name }}
+                          </button>
+                          <ToolCallBadge
+                            v-else
+                            :kind="
+                              classifyToolCall(
+                                toolCall.function.name,
+                                toolCall.function.arguments,
+                                getToolResultText(toolCall.id),
+                              )
+                            "
+                            @click="
+                              openToolPreview(
+                                classifyToolCall(
+                                  toolCall.function.name,
+                                  toolCall.function.arguments,
+                                  getToolResultText(toolCall.id),
+                                ),
+                              )
+                            "
+                          />
+                        </template>
+                      </div>
+                      <div
+                        v-if="activeToolDetail?.messageId === message.id"
+                        class="chat-tool-detail-inline"
+                      >
+                        <div class="chat-tool-detail-inline-header">
+                          <strong>{{ activeToolDetail.toolCall.function.name }}</strong>
+                          <button class="chat-tool-detail-inline-close" @click="closeToolDetail">
+                            收起
+                          </button>
+                        </div>
+                        <div class="chat-tool-detail-inline-block">
+                          <div class="chat-tool-detail-caption">tool_call_id</div>
+                          <code>{{ activeToolDetail.toolCall.id }}</code>
+                        </div>
+                        <div class="chat-tool-detail-inline-block">
+                          <div class="chat-tool-detail-caption">arguments</div>
+                          <pre>{{ formatToolPayload(activeToolDetail.toolCall.function.arguments) }}</pre>
+                        </div>
+                        <div class="chat-tool-detail-inline-block">
+                          <div class="chat-tool-detail-caption">result</div>
+                          <pre>{{ activeToolDetail.result || "(空结果)" }}</pre>
+                        </div>
+                      </div>
+                    </div>
+                  </div>
+                </div>
+                <div v-if="group.role !== 'assistant'" class="chat-bubble-col">
+                  <div
+                    v-for="(message, idx) in group.messages"
+                    :key="message.id + '-' + idx"
+                    class="chat-bubble"
+                    :class="message.role"
+                  >
+                    <div
+                      class="chat-bubble-content markdown-body"
+                      v-html="renderMessageContent(message.content, message.streaming)"
+                    ></div>
+                    <div class="chat-bubble-time">{{ formatChatTime(message.timestamp) }}</div>
+                  </div>
+                </div>
+              </div>
+            </div>
+
+            <div class="chat-input-area">
+              <div v-if="!isChatEligible" class="chat-not-supported">
+                <div class="chat-not-supported-icon">🚫</div>
+                <div class="chat-not-supported-title">此 Agent 不支持在 Dashboard 聊天</div>
+                <div class="chat-not-supported-desc">请在 QQ 群或 HTTP Stream 端点中使用该 Agent。</div>
+              </div>
+              <template v-else>
+                <div v-if="pendingAskUser" class="ask-user-panel">
+                  <div class="ask-user-question">{{ pendingAskUser.question }}</div>
+                  <div v-if="pendingAskUser.details" class="ask-user-details">
+                    {{ pendingAskUser.details }}
+                  </div>
+                  <div class="ask-user-row">
+                    <input
+                      v-model="askUserAnswer"
+                      type="text"
+                      :placeholder="pendingAskUser.placeholder || '请输入补充信息'"
+                      @input="clearChatError"
+                      @keydown.enter.prevent="submitAskUserAnswer"
+                    />
+                    <button
+                      class="btn primary"
+                      :disabled="!canSubmitAskUser"
+                      @click="submitAskUserAnswer"
+                    >
+                      提交补充信息
+                    </button>
+                  </div>
+                </div>
+                <textarea
+                  v-model="draftMessage"
+                  placeholder="输入消息"
+                  @keydown.ctrl.enter.prevent="sendMessage"
+                  @input="clearChatError"
+                />
+                <div class="chat-input-actions">
+                  <button class="btn ghost" @click="startNewSession">新对话</button>
+                  <div class="chat-input-right">
+                    <div v-if="isChatEligible" class="chat-model-bar">
+                      <div class="model-picker" :class="{ open: openPicker === 'model' }">
+                        <button
+                          class="model-chip"
+                          @click.stop="openPicker = openPicker === 'model' ? null : 'model'"
+                        >
+                          {{ selectedModelLabel }}
+                          <svg
+                            class="chip-chevron"
+                            viewBox="0 0 24 24"
+                            fill="none"
+                            stroke="currentColor"
+                            stroke-width="2.5"
+                            stroke-linecap="round"
+                            stroke-linejoin="round"
+                          >
+                            <polyline points="6 9 12 15 18 9" />
+                          </svg>
+                        </button>
+                        <div v-if="openPicker === 'model'" class="model-picker-dropdown">
+                          <button
+                            class="model-picker-item"
+                            :class="{ active: selectedModelId === '' }"
+                            @click.stop="selectModel('')"
+                          >
+                            默认模型
+                          </button>
+                          <button
+                            v-for="model in chatModels"
+                            :key="model.config_id"
+                            class="model-picker-item"
+                            :class="{ active: selectedModelId === model.config_id }"
+                            @click.stop="selectModel(model.config_id)"
+                          >
+                            {{ model.name }}
+                          </button>
+                        </div>
+                      </div>
+
+                      <div class="model-picker" :class="{ open: openPicker === 'thinking' }">
+                        <button
+                          class="model-chip"
+                          @click.stop="openPicker = openPicker === 'thinking' ? null : 'thinking'"
+                        >
+                          {{ selectedThinkingLabel }}
+                          <svg
+                            class="chip-chevron"
+                            viewBox="0 0 24 24"
+                            fill="none"
+                            stroke="currentColor"
+                            stroke-width="2.5"
+                            stroke-linecap="round"
+                            stroke-linejoin="round"
+                          >
+                            <polyline points="6 9 12 15 18 9" />
+                          </svg>
+                        </button>
+                        <div v-if="openPicker === 'thinking'" class="model-picker-dropdown">
+                          <button
+                            class="model-picker-item"
+                            :class="{ active: selectedThinkingType === '' }"
+                            @click.stop="selectThinkingType('')"
+                          >
+                            默认{{ selectedModelLlmConfig?.thinking_type ? (selectedModelLlmConfig.thinking_type === 'enabled' ? '(启用)' : '(禁用)') : '' }}
+                          </button>
+                          <button
+                            class="model-picker-item"
+                            :class="{ active: selectedThinkingType === 'enabled' }"
+                            @click.stop="selectThinkingType('enabled')"
+                          >
+                            启用
+                          </button>
+                          <button
+                            class="model-picker-item"
+                            :class="{ active: selectedThinkingType === 'disabled' }"
+                            @click.stop="selectThinkingType('disabled')"
+                          >
+                            禁用
+                          </button>
+                        </div>
+                      </div>
+
+                      <div class="model-picker" :class="{ open: openPicker === 'effort' }">
+                        <button
+                          class="model-chip"
+                          @click.stop="openPicker = openPicker === 'effort' ? null : 'effort'"
+                        >
+                          {{ selectedEffortLabel }}
+                          <svg
+                            class="chip-chevron"
+                            viewBox="0 0 24 24"
+                            fill="none"
+                            stroke="currentColor"
+                            stroke-width="2.5"
+                            stroke-linecap="round"
+                            stroke-linejoin="round"
+                          >
+                            <polyline points="6 9 12 15 18 9" />
+                          </svg>
+                        </button>
+                        <div v-if="openPicker === 'effort'" class="model-picker-dropdown">
+                          <button
+                            class="model-picker-item"
+                            :class="{ active: selectedReasoningEffort === '' }"
+                            @click.stop="selectReasoningEffort('')"
+                          >
+                            默认{{ selectedModelLlmConfig?.reasoning_effort ? `(${selectedModelLlmConfig.reasoning_effort})` : '' }}
+                          </button>
+                          <button
+                            class="model-picker-item"
+                            :class="{ active: selectedReasoningEffort === 'low' }"
+                            @click.stop="selectReasoningEffort('low')"
+                          >
+                            low
+                          </button>
+                          <button
+                            class="model-picker-item"
+                            :class="{ active: selectedReasoningEffort === 'medium' }"
+                            @click.stop="selectReasoningEffort('medium')"
+                          >
+                            medium
+                          </button>
+                          <button
+                            class="model-picker-item"
+                            :class="{ active: selectedReasoningEffort === 'high' }"
+                            @click.stop="selectReasoningEffort('high')"
+                          >
+                            high
+                          </button>
+                          <button
+                            class="model-picker-item"
+                            :class="{ active: selectedReasoningEffort === 'max' }"
+                            @click.stop="selectReasoningEffort('max')"
+                          >
+                            max
+                          </button>
+                        </div>
+                      </div>
+
+                      <div class="model-settings" :class="{ open: openPicker === 'settings' }">
+                        <button
+                          class="model-chip icon-only"
+                          title="模型设置"
+                          @click.stop="openPicker = openPicker === 'settings' ? null : 'settings'"
+                        >
+                          <svg
+                            viewBox="0 0 24 24"
+                            fill="none"
+                            stroke="currentColor"
+                            stroke-width="2"
+                            stroke-linecap="round"
+                            stroke-linejoin="round"
+                          >
+                            <circle cx="12" cy="12" r="3" />
+                            <path
+                              d="M19.4 15a1.65 1.65 0 0 0 .33 1.82l.06.06a2 2 0 0 1 0 2.83 2 2 0 0 1-2.83 0l-.06-.06a1.65 1.65 0 0 0-1.82-.33 1.65 1.65 0 0 0-1 1.51V21a2 2 0 0 1-2 2 2 2 0 0 1-2-2v-.09A1.65 1.65 0 0 0 9 19.4a1.65 1.65 0 0 0-1.82.33l-.06.06a2 2 0 0 1-2.83 0 2 2 0 0 1 0-2.83l.06-.06a1.65 1.65 0 0 0 .33-1.82 1.65 1.65 0 0 0-1.51-1H3a2 2 0 0 1-2-2 2 2 0 0 1 2-2h.09A1.65 1.65 0 0 0 4.6 9a1.65 1.65 0 0 0-.33-1.82l-.06-.06a2 2 0 0 1 0-2.83 2 2 0 0 1 2.83 0l.06.06a1.65 1.65 0 0 0 1.82.33H9a1.65 1.65 0 0 0 1-1.51V3a2 2 0 0 1 2-2 2 2 0 0 1 2 2v.09a1.65 1.65 0 0 0 1 1.51 1.65 1.65 0 0 0 1.82-.33l.06-.06a2 2 0 0 1 2.83 0 2 2 0 0 1 0 2.83l-.06.06a1.65 1.65 0 0 0-.33 1.82V9a1.65 1.65 0 0 0 1.51 1H21a2 2 0 0 1 2 2 2 2 0 0 1-2 2h-.09a1.65 1.65 0 0 0-1.51 1z"
+                            />
+                          </svg>
+                        </button>
+                      </div>
+                    </div>
+                    <button class="btn primary" :disabled="sending || !canSend" @click="sendMessage">
+                      {{ sending ? "推理中..." : "发送" }}
+                    </button>
+                  </div>
+                </div>
+              </template>
+              <div v-if="chatErrorMessage" class="chat-error-box" role="alert">
+                {{ chatErrorMessage }}
+              </div>
+            </div>
+          </div>
+        </div>
+      </section>
+    </div>
+
+    <Teleport to="body">
+      <div
+        v-if="toolPreviewState"
+        class="tool-preview-overlay"
+        @click.self="closeToolPreview"
+        @keydown="handleToolPreviewKeydown"
+      >
+        <div class="tool-preview-panel">
+          <div class="tool-preview-header">
+            <template v-if="toolPreviewState.kind.type === 'create_file'">
+              <span class="badge-icon">📄</span> 创建文件: {{ toolPreviewState.kind.filename }}
+            </template>
+            <template v-else-if="toolPreviewState.kind.type === 'delete_file'">
+              <span class="badge-icon">🗑️</span> 删除文件: {{ toolPreviewState.kind.filename }}
+            </template>
+            <template v-else-if="toolPreviewState.kind.type === 'edit_file'">
+              <span class="badge-icon">✏️</span> 编辑文件: {{ toolPreviewState.kind.filename }}
+            </template>
+            <template v-else-if="toolPreviewState.kind.type === 'exec_cmd'">
+              <span class="cmd-prefix">&gt;</span> {{ toolPreviewState.kind.command }}
+            </template>
+            <button class="tool-preview-close" @click="closeToolPreview">✕</button>
+          </div>
+          <div class="tool-preview-body">
+            <template v-if="toolPreviewState.kind.type === 'create_file'">
+              <pre class="tool-preview-code tool-preview-code--create">
+                {{ toolPreviewState.kind.content }}
+              </pre>
+            </template>
+            <template v-else-if="toolPreviewState.kind.type === 'delete_file'">
+              <div class="tool-preview-info tool-preview-info--delete">
+                <p>已删除文件: <code>{{ toolPreviewState.kind.filename }}</code></p>
+                <p v-if="toolPreviewState.kind.lineCount != null">
+                  共 {{ toolPreviewState.kind.lineCount }} 行
+                </p>
+              </div>
+            </template>
+            <template v-else-if="toolPreviewState.kind.type === 'edit_file'">
+              <div class="tool-preview-diff">
+                <div
+                  v-for="(hunk, idx) in editHunks(toolPreviewState.kind.edits)"
+                  :key="idx"
+                  class="tool-preview-hunk"
+                >
+                  <div
+                    v-for="line in hunk.removed"
+                    :key="line"
+                    class="tool-preview-diff-line tool-preview-diff-line--removed"
+                  >
+                    <span class="diff-marker">-</span> {{ line }}
+                  </div>
+                  <div
+                    v-for="line in hunk.added"
+                    :key="line"
+                    class="tool-preview-diff-line tool-preview-diff-line--added"
+                  >
+                    <span class="diff-marker">+</span> {{ line }}
+                  </div>
+                </div>
+              </div>
+            </template>
+            <template v-else-if="toolPreviewState.kind.type === 'exec_cmd'">
+              <template v-if="toolPreviewState.kind.hasResult">
+                <div v-if="toolPreviewState.kind.stdout" class="tool-preview-output">
+                  <div class="tool-preview-output-label">stdout</div>
+                  <pre class="tool-preview-code tool-preview-code--cmd">
+                    {{ toolPreviewState.kind.stdout }}
+                  </pre>
+                </div>
+                <div v-if="toolPreviewState.kind.stderr" class="tool-preview-output">
+                  <div class="tool-preview-output-label tool-preview-output-label--error">
+                    stderr
+                  </div>
+                  <pre class="tool-preview-code tool-preview-code--cmd tool-preview-code--error">
+                    {{ toolPreviewState.kind.stderr }}
+                  </pre>
+                </div>
+              </template>
+              <div v-else class="tool-preview-no-result">无结果</div>
+            </template>
+          </div>
+        </div>
+      </div>
+    </Teleport>
+  </section>
+</template>
+
+<script setup lang="ts">
+import { computed, nextTick, onMounted, onUnmounted, reactive, ref, watch } from "vue";
+import MarkdownIt from "markdown-it";
+
+import {
+  chat,
+  system,
+  type ServiceWithRuntime,
+  type ChatHistoryRecord,
+  type ChatToolCall,
+  type ChatSessionSummary,
+  type ChatStreamEvent,
+  type LlmConfig,
+} from "../../api/client";
+import { formatTime, agentAvatarUrl, agentInitial, getAvatarDisplayUrl } from "../model";
+import ToolCallBadge from "./ToolCallBadge.vue";
+
+type ChatRole = "user" | "assistant" | "tool";
+type LiveToolCall = {
+  call_id: string;
+  name: string;
+  arguments: unknown;
+  result?: string;
+  done: boolean;
+};
+type ChatMessage = {
+  id: string;
+  role: ChatRole;
+  content: string;
+  thinkingContent?: string;
+  thinkingExpanded?: boolean;
+  streaming?: boolean;
+  timestamp?: string;
+  toolCalls: ChatToolCall[];
+  toolCallId?: string | null;
+  linkedToolCall?: ChatToolCall | null;
+  agentAvatarUrl?: string;
+  agentName?: string;
+  liveToolCalls?: LiveToolCall[];
+};
+type ToolDetail = {
+  messageId: string;
+  toolCall: ChatToolCall;
+  result: string;
+};
+type LineEditSpec = {
+  start_line: number;
+  end_line: number;
+  replacement_lines: string[];
+};
+type ToolCallKind =
+  | { type: "create_file"; filename: string; lineCount: number; content: string }
+  | { type: "delete_file"; filename: string; lineCount: number | null }
+  | { type: "edit_file"; filename: string; addedLines: number; removedLines: number; edits: LineEditSpec[] }
+  | { type: "exec_cmd"; command: string; hasResult: boolean; stdout?: string; stderr?: string }
+  | { type: "generic"; name: string };
+
+const props = defineProps<{
+  agentId?: string;
+  sessionId?: string;
+  embedded?: boolean;
+}>();
+
+const emit = defineEmits<{
+  (e: "update:sessionId", sessionId: string): void;
+}>();
+
+function basename(p: string): string {
+  const seg = p.replace(/\\/g, "/").split("/");
+  return seg[seg.length - 1] || p;
+}
+
+function safeParseJson<T>(raw: unknown): T | null {
+  if (raw == null) return null;
+  try {
+    return typeof raw === "string" ? (JSON.parse(raw) as T) : (raw as T);
+  } catch {
+    return null;
+  }
+}
+
+function classifyToolCall(name: string, arguments_: unknown, result?: string): ToolCallKind {
+  if (name === "create_file") {
+    const args = safeParseJson<{ path?: string; content?: string }>(arguments_);
+    if (args?.path != null && args?.content != null) {
+      return {
+        type: "create_file",
+        filename: basename(args.path),
+        lineCount: args.content.split("\n").length,
+        content: args.content,
+      };
+    }
+  }
+  if (name === "delete_file") {
+    const args = safeParseJson<{ path?: string }>(arguments_);
+    const res = safeParseJson<{ line_count?: number | null }>(result);
+    if (args?.path != null) {
+      return {
+        type: "delete_file",
+        filename: basename(args.path),
+        lineCount: res?.line_count ?? null,
+      };
+    }
+  }
+  if (name === "edit_file") {
+    const args = safeParseJson<{ path?: string; edits?: LineEditSpec[] }>(arguments_);
+    if (args?.path != null && Array.isArray(args.edits)) {
+      let addedLines = 0;
+      let removedLines = 0;
+      for (const edit of args.edits) {
+        removedLines += Math.max(0, edit.end_line - edit.start_line + 1);
+        addedLines += edit.replacement_lines.length;
+      }
+      return {
+        type: "edit_file",
+        filename: basename(args.path),
+        addedLines,
+        removedLines,
+        edits: args.edits,
+      };
+    }
+  }
+  if (name === "exec_cmd") {
+    const args = safeParseJson<{ command?: string }>(arguments_);
+    const res = safeParseJson<{ stdout?: string; stderr?: string }>(result);
+    const stdout = res?.stdout ?? "";
+    const stderr = res?.stderr ?? "";
+    const hasResult = stdout.length > 0 || stderr.length > 0;
+    if (args?.command != null) {
+      return {
+        type: "exec_cmd",
+        command: args.command,
+        hasResult,
+        stdout: hasResult ? stdout : undefined,
+        stderr: hasResult ? stderr : undefined,
+      };
+    }
+  }
+  return { type: "generic", name };
+}
+
+type PendingNewConversationCommand = {
+  passthroughText: string | null;
+};
+type PendingAskUser = {
+  question: string;
+  details?: string;
+  placeholder?: string;
+};
+type StreamState = {
+  assistantMessageId: string | null;
+  pendingNewConversation: PendingNewConversationCommand | null;
+  requestText: string;
+};
+
+const services = ref<ServiceWithRuntime[]>([]);
+const servicesLoading = ref(false);
+const sessions = ref<ChatSessionSummary[]>([]);
+const activeSessionId = ref("");
+const selectedServiceId = ref("");
+const draftMessage = ref("");
+const workspacePath = ref("");
+const pickingDirectory = ref(false);
+const sending = ref(false);
+const chatErrorMessage = ref("");
+const messagesContainer = ref<HTMLElement | null>(null);
+const messages = ref<ChatMessage[]>([]);
+const activeToolCallId = ref("");
+const expandedLiveToolCalls = ref(new Set<string>());
+const llmModels = ref<LlmConfig[]>([]);
+const selectedModelId = ref("");
+const selectedThinkingType = ref<"" | "enabled" | "disabled">("");
+const selectedReasoningEffort = ref<"" | "low" | "medium" | "high" | "max">("");
+const openPicker = ref<"model" | "thinking" | "effort" | "settings" | null>(null);
+const stats = reactive({
+  connections: 0,
+  llm: 0,
+  services: 0,
+});
+const markdown = new MarkdownIt({
+  html: false,
+  breaks: true,
+  linkify: true,
+});
+
+const chatEligibleServiceTypes = new Set(["http_stream", "workspace"]);
+const selectedService = computed(
+  () => services.value.find((agent) => agent.config_id === selectedServiceId.value) ?? null,
+);
+const selectedServiceType = computed(() => selectedService.value?.agent_type?.type ?? "");
+const isChatEligible = computed(() => chatEligibleServiceTypes.has(selectedServiceType.value));
+const isWorkspaceService = computed(() => selectedServiceType.value === "workspace");
+const groupedSessions = computed(() => {
+  const groups = new Map<string, ChatSessionSummary[]>();
+  for (const session of sessions.value) {
+    const key = session.workspace_path ?? "__default__";
+    if (!groups.has(key)) groups.set(key, []);
+    groups.get(key)!.push(session);
+  }
+  const result: Array<{
+    pathKey: string;
+    path: string | null;
+    label: string;
+    sessions: ChatSessionSummary[];
+  }> = [];
+  for (const [key, items] of groups) {
+    items.sort((a, b) => b.updated_at.localeCompare(a.updated_at));
+    result.push({
+      pathKey: key,
+      path: key === "__default__" ? null : key,
+      label: key === "__default__" ? "默认路径" : key,
+      sessions: items,
+    });
+  }
+  result.sort((a, b) => b.sessions[0].updated_at.localeCompare(a.sessions[0].updated_at));
+  return result;
+});
+const chatModels = computed(() => llmModels.value.filter((item) => item.model.type === "chat_llm"));
+const selectedModelLlmConfig = computed(() => {
+  const modelId = selectedModelId.value || defaultAgentModelId.value;
+  const model = chatModels.value.find((m) => m.config_id === modelId);
+  if (model?.model.type === "chat_llm") {
+    return model.model.llm;
+  }
+  return null;
+});
+const defaultAgentModelId = computed(() => {
+  const agent = selectedService.value;
+  if (!agent) {
+    return "";
+  }
+  const agentType = agent.agent_type as Record<string, unknown>;
+  return String(agentType.llm_ref_id ?? "");
+});
+const selectedModelLabel = computed(() => {
+  if (!selectedModelId.value) {
+    return "默认模型";
+  }
+  const model = chatModels.value.find((m) => m.config_id === selectedModelId.value);
+  return model?.name ?? "默认模型";
+});
+const selectedThinkingLabel = computed(() => {
+  const defaultType = selectedModelLlmConfig.value?.thinking_type;
+  const defaultLabel = defaultType ? (defaultType === "enabled" ? "(启用)" : "(禁用)") : "";
+  if (!selectedThinkingType.value) {
+    return `默认${defaultLabel}`;
+  }
+  return selectedThinkingType.value === "enabled" ? "启用" : "禁用";
+});
+const selectedEffortLabel = computed(() => {
+  const defaultEffort = selectedModelLlmConfig.value?.reasoning_effort;
+  const defaultLabel = defaultEffort ? `(${defaultEffort})` : "";
+  if (!selectedReasoningEffort.value) {
+    return `默认${defaultLabel}`;
+  }
+  return selectedReasoningEffort.value;
+});
+const canSend = computed(() =>
+  !!selectedService.value &&
+  isChatEligible.value &&
+  selectedService.value.runtime.status === "running" &&
+  draftMessage.value.trim().length > 0,
+);
+const selectedAgentAvatarUrl = computed(() => agentAvatarUrl(selectedService.value));
+const selectedAgentAvatarFallback = computed(() => {
+  const name = selectedService.value?.name ?? "Bot";
+  return agentInitial(name);
+});
+const pendingAskUser = ref<PendingAskUser | null>(null);
+const askUserAnswer = ref("");
+const canSubmitAskUser = computed(() =>
+  isChatEligible.value &&
+  isWorkspaceService.value &&
+  !!pendingAskUser.value &&
+  selectedService.value?.runtime.status === "running" &&
+  askUserAnswer.value.trim().length > 0 &&
+  !sending.value,
+);
+
+function parseNewConversationCommand(input: string): PendingNewConversationCommand | null {
+  const match = input.trim().match(/^\/(new|clear|reset)(?:\s+([\s\S]*))?$/i);
+  if (!match) {
+    return null;
+  }
+
+  const passthroughText = (match[2] ?? "").trim();
+  return {
+    passthroughText: passthroughText.length > 0 ? passthroughText : null,
+  };
+}
+
+function messageAvatarUrl(record: ChatHistoryRecord): string {
+  if (record.agent_avatar_url) {
+    return getAvatarDisplayUrl(record.agent_avatar_url);
+  }
+  const agent = services.value.find((a) => a.config_id === record.agent_id);
+  return agentAvatarUrl(agent);
+}
+
+type MessageGroup = {
+  id: string;
+  role: ChatRole;
+  messages: ChatMessage[];
+  avatarUrl?: string;
+  agentName?: string;
+};
+
+const messageGroups = computed(() => {
+  const filtered = messages.value.filter((m) => m.role !== "tool");
+  const groups: MessageGroup[] = [];
+  let currentGroup: MessageGroup | null = null;
+
+  for (const message of filtered) {
+    if (currentGroup && currentGroup.role === message.role) {
+      currentGroup.messages.push(message);
+    } else {
+      currentGroup = {
+        id: `group-${message.id}`,
+        role: message.role,
+        messages: [message],
+        avatarUrl:
+          message.role === "assistant"
+            ? message.agentAvatarUrl || selectedAgentAvatarUrl.value || undefined
+            : undefined,
+        agentName: message.agentName || selectedService.value?.name,
+      };
+      groups.push(currentGroup);
+    }
+  }
+  return groups;
+});
+
+const activeToolDetail = computed<ToolDetail | null>(() => {
+  if (!activeToolCallId.value) {
+    return null;
+  }
+  for (const message of messages.value) {
+    const toolCall = message.toolCalls.find((item) => item.id === activeToolCallId.value);
+    if (!toolCall) {
+      continue;
+    }
+    const resultMessage = messages.value.find(
+      (item) => item.role === "tool" && item.toolCallId === toolCall.id,
+    );
+    return {
+      messageId: message.id,
+      toolCall,
+      result: resultMessage?.content ?? "",
+    };
+  }
+  return null;
+});
+
+function readableAgentType(type: string): string {
+  if (type === "http_stream") {
+    return "HTTP stream service";
+  }
+  if (type === "workspace") {
+    return "Workspace Agent Service";
+  }
+  return "QQ Chat Agent Service";
+}
+
+function toApiMessages() {
+  return messages.value
+    .filter((item) => item.content.trim().length > 0 || item.toolCalls.length > 0 || !!item.toolCallId)
+    .map((item) => ({
+      role: item.role,
+      content: item.content,
+      tool_calls: item.toolCalls.length > 0 ? item.toolCalls : undefined,
+      tool_call_id: item.toolCallId ?? undefined,
+    }));
+}
+
+function applyHistory(records: ChatHistoryRecord[]) {
+  const mapped: ChatMessage[] = records
+    .filter((item) => item.role === "user" || item.role === "assistant" || item.role === "tool")
+    .map((item) => ({
+      id: item.message_id,
+      role: item.role as ChatRole,
+      content: item.content,
+      thinkingContent: item.reasoning_content ?? undefined,
+      thinkingExpanded: !!item.reasoning_content,
+      timestamp: item.timestamp,
+      toolCalls: item.tool_calls ?? [],
+      toolCallId: item.tool_call_id ?? null,
+      linkedToolCall: null,
+      agentAvatarUrl: messageAvatarUrl(item) || undefined,
+      agentName: item.agent_name || undefined,
+    }));
+  const toolCallMap = new Map<string, ChatToolCall>();
+  for (const message of mapped) {
+    for (const toolCall of message.toolCalls) {
+      toolCallMap.set(toolCall.id, toolCall);
+    }
+  }
+  for (const message of mapped) {
+    if (message.role === "tool" && message.toolCallId) {
+      message.linkedToolCall = toolCallMap.get(message.toolCallId) ?? null;
+    }
+  }
+  messages.value = mapped;
+  if (activeToolCallId.value) {
+    const stillExists = mapped.some((message) =>
+      message.toolCalls.some((toolCall) => toolCall.id === activeToolCallId.value),
+    );
+    if (!stillExists) {
+      activeToolCallId.value = "";
+    }
+  }
+  scrollToBottom();
+}
+
+function openToolDetail(messageId: string, toolCallId: string) {
+  const message = messages.value.find((item) => item.id === messageId);
+  if (!message || !message.toolCalls.some((item) => item.id === toolCallId)) {
+    return;
+  }
+  activeToolCallId.value = activeToolCallId.value === toolCallId ? "" : toolCallId;
+}
+
+function closeToolDetail() {
+  activeToolCallId.value = "";
+}
+
+function getToolResultText(toolCallId: string): string | undefined {
+  const resultMessage = messages.value.find(
+    (item) => item.role === "tool" && item.toolCallId === toolCallId,
+  );
+  return resultMessage?.content || undefined;
+}
+
+type ToolPreviewData = {
+  kind: ToolCallKind;
+  toolCallId: string;
+};
+
+const toolPreviewState = ref<ToolPreviewData | null>(null);
+
+function openToolPreview(kind: ToolCallKind) {
+  toolPreviewState.value = { kind, toolCallId: "" };
+}
+
+function closeToolPreview() {
+  toolPreviewState.value = null;
+}
+
+function handleToolPreviewKeydown(e: KeyboardEvent) {
+  if (e.key === "Escape" && toolPreviewState.value) {
+    closeToolPreview();
+  }
+}
+
+function editHunks(edits: LineEditSpec[]): { startLine: number; removed: string[]; added: string[] }[] {
+  const sorted = [...edits].sort((a, b) => a.start_line - b.start_line);
+  const hunks: { startLine: number; removed: string[]; added: string[] }[] = [];
+  for (const edit of sorted) {
+    const removedCount = Math.max(0, edit.end_line - edit.start_line + 1);
+    hunks.push({
+      startLine: edit.start_line,
+      removed: Array.from({ length: removedCount }, (_, i) => `L${edit.start_line + i}`),
+      added: edit.replacement_lines.map(
+        (line, i) => `L${edit.start_line + i}` + (line.length > 0 ? ": " + line : ""),
+      ),
+    });
+  }
+  return hunks;
+}
+
+function toggleLiveToolCall(callId: string) {
+  if (expandedLiveToolCalls.value.has(callId)) {
+    expandedLiveToolCalls.value.delete(callId);
+  } else {
+    expandedLiveToolCalls.value.add(callId);
+  }
+  expandedLiveToolCalls.value = new Set(expandedLiveToolCalls.value);
+}
+
+function formatToolPayload(payload: unknown): string {
+  if (payload == null) {
+    return "null";
+  }
+  if (typeof payload === "string") {
+    return payload;
+  }
+  try {
+    return JSON.stringify(payload, null, 2);
+  } catch {
+    return String(payload);
+  }
+}
+
+function formatChatTime(timestamp?: string): string {
+  if (!timestamp) {
+    return "";
+  }
+  const date = new Date(timestamp);
+  if (Number.isNaN(date.getTime())) {
+    return timestamp;
+  }
+  return date.toLocaleTimeString("zh-CN", {
+    hour: "2-digit",
+    minute: "2-digit",
+    second: "2-digit",
+    hour12: false,
+  });
+}
+
+function renderMessageContent(content: string, streaming = false): string {
+  const text = content || (streaming ? "..." : "");
+  return markdown.render(text);
+}
+
+function scrollToBottom() {
+  nextTick(() => {
+    if (messagesContainer.value) {
+      messagesContainer.value.scrollTop = messagesContainer.value.scrollHeight;
+    }
+  });
+}
+
+function clearChatError() {
+  chatErrorMessage.value = "";
+}
+
+function clearPendingAskUser() {
+  pendingAskUser.value = null;
+  askUserAnswer.value = "";
+}
+
+function pruneFailedAssistantPlaceholder(assistantMessageId: string | null) {
+  if (!assistantMessageId) {
+    return;
+  }
+  const index = messages.value.findIndex((item) => item.id === assistantMessageId);
+  if (index < 0) {
+    return;
+  }
+
+  const message = messages.value[index];
+  message.streaming = false;
+
+  const hasVisibleContent =
+    message.content.trim().length > 0 || (message.thinkingContent?.trim().length ?? 0) > 0;
+  const hasToolActivity =
+    (message.liveToolCalls?.length ?? 0) > 0 || message.toolCalls.length > 0;
+  if (!hasVisibleContent && !hasToolActivity) {
+    messages.value.splice(index, 1);
+  }
+}
+
+function applyInferenceFailure(streamState: StreamState, errorMessage: string) {
+  pruneFailedAssistantPlaceholder(streamState.assistantMessageId);
+  chatErrorMessage.value = `推理失败: ${errorMessage}`;
+  if (!draftMessage.value.trim()) {
+    draftMessage.value = streamState.requestText;
+  }
+}
+
+async function reloadSessions() {
+  const result = await chat.listSessions(selectedServiceId.value || undefined);
+  sessions.value = result.sessions;
+}
+
+async function openSession(sessionId: string) {
+  activeSessionId.value = sessionId;
+  emit("update:sessionId", sessionId);
+  clearChatError();
+  clearPendingAskUser();
+  const result = await chat.getSessionMessages(sessionId);
+  const firstRecord = result.messages[0];
+  if (firstRecord?.agent_id && services.value.some((a) => a.config_id === firstRecord.agent_id)) {
+    selectedServiceId.value = firstRecord.agent_id;
+  }
+  workspacePath.value =
+    result.messages[result.messages.length - 1]?.workspace_path ??
+    firstRecord?.workspace_path ??
+    workspacePath.value;
+  const latestRecord = result.messages[result.messages.length - 1];
+  if (latestRecord?.pending_ask_user?.question) {
+    pendingAskUser.value = {
+      question: latestRecord.pending_ask_user.question,
+      details: latestRecord.pending_ask_user.details ?? undefined,
+      placeholder: latestRecord.pending_ask_user.placeholder ?? undefined,
+    };
+  }
+  applyHistory(result.messages);
+}
+
+async function pickDirectory() {
+  pickingDirectory.value = true;
+  try {
+    const result = await system.selectDirectory();
+    if (result.path) {
+      workspacePath.value = result.path;
+    }
+  } catch (error) {
+    chatErrorMessage.value = `选择目录失败: ${(error as Error).message}`;
+  } finally {
+    pickingDirectory.value = false;
+  }
+}
+
+function startNewSession() {
+  activeSessionId.value = "";
+  emit("update:sessionId", "");
+  messages.value = [];
+  activeToolCallId.value = "";
+  expandedLiveToolCalls.value = new Set();
+  clearChatError();
+  clearPendingAskUser();
+}
+
+function selectModel(id: string) {
+  selectedModelId.value = id;
+  openPicker.value = null;
+}
+
+function selectThinkingType(value: "" | "enabled" | "disabled") {
+  selectedThinkingType.value = value;
+  openPicker.value = null;
+}
+
+function selectReasoningEffort(value: "" | "low" | "medium" | "high" | "max") {
+  selectedReasoningEffort.value = value;
+  openPicker.value = null;
+}
+
+function closePickersOnClickOutside(event: MouseEvent) {
+  const target = event.target as HTMLElement;
+  if (!target.closest(".model-picker") && !target.closest(".model-settings")) {
+    openPicker.value = null;
+  }
+}
+
+watch(selectedServiceId, async () => {
+  await reloadSessions();
+  startNewSession();
+  selectedModelId.value = defaultAgentModelId.value;
+  selectedThinkingType.value = "";
+  selectedReasoningEffort.value = "";
+  if (!isWorkspaceService.value) {
+    workspacePath.value = "";
+  }
+  if (!isChatEligible.value) {
+    clearPendingAskUser();
+  }
+});
+
+watch(selectedModelId, () => {
+  selectedThinkingType.value = "";
+  selectedReasoningEffort.value = "";
+});
+
+async function removeSession(sessionId: string) {
+  if (!confirm("确定要删除该会话吗？此操作不可恢复。")) {
+    return;
+  }
+  try {
+    await chat.deleteSession(sessionId);
+    sessions.value = sessions.value.filter((s) => s.session_id !== sessionId);
+    if (activeSessionId.value === sessionId) {
+      startNewSession();
+    }
+  } catch (error) {
+    alert(`删除失败: ${(error as Error).message}`);
+  }
+}
+
+function applyStreamEvent(event: ChatStreamEvent, streamState: StreamState) {
+  if (event.type === "error") {
+    applyInferenceFailure(streamState, event.error ?? "未知错误");
+    return;
+  }
+
+  if (event.type === "ask_user" && event.question) {
+    pendingAskUser.value = {
+      question: event.question,
+      details: event.details ?? undefined,
+      placeholder: event.placeholder ?? undefined,
+    };
+    askUserAnswer.value = "";
+    return;
+  }
+
+  if (event.type === "start") {
+    if (streamState.pendingNewConversation) {
+      startNewSession();
+      if (event.session_id) {
+        activeSessionId.value = event.session_id;
+        emit("update:sessionId", event.session_id);
+      }
+
+      const passthroughText = streamState.pendingNewConversation.passthroughText;
+      if (passthroughText) {
+        messages.value.push({
+          id: `local-user-${crypto.randomUUID()}`,
+          role: "user",
+          content: passthroughText,
+          timestamp: new Date().toISOString(),
+          toolCalls: [],
+          toolCallId: null,
+          linkedToolCall: null,
+        });
+
+        const assistantMessageId = event.message_id ?? `local-assistant-${crypto.randomUUID()}`;
+        messages.value.push({
+          id: assistantMessageId,
+          role: "assistant",
+          content: "",
+          streaming: true,
+          timestamp: new Date().toISOString(),
+          toolCalls: [],
+          toolCallId: null,
+          linkedToolCall: null,
+        });
+        streamState.assistantMessageId = assistantMessageId;
+      } else {
+        streamState.assistantMessageId = event.message_id ?? null;
+      }
+
+      streamState.pendingNewConversation = null;
+      scrollToBottom();
+      return;
+    }
+
+    if (event.session_id) {
+      activeSessionId.value = event.session_id;
+      emit("update:sessionId", event.session_id);
+    }
+    if (event.message_id) {
+      const currentAssistantId = streamState.assistantMessageId;
+      const message = currentAssistantId
+        ? messages.value.find((item) => item.id === currentAssistantId || item.id === event.message_id)
+        : undefined;
+      if (message) {
+        message.id = event.message_id;
+      }
+      streamState.assistantMessageId = event.message_id;
+    }
+  }
+
+  if (event.type === "delta") {
+    const targetId = event.message_id || streamState.assistantMessageId;
+    if (!targetId) {
+      return;
+    }
+    const message = messages.value.find((item) => item.id === targetId);
+    if (message) {
+      message.content += event.token ?? "";
+      message.streaming = true;
+      scrollToBottom();
+    }
+  }
+
+  if (event.type === "thinking_delta") {
+    const targetId = event.message_id || streamState.assistantMessageId;
+    if (!targetId) {
+      return;
+    }
+    const message = messages.value.find((item) => item.id === targetId);
+    if (message) {
+      if (!message.thinkingContent) {
+        message.thinkingContent = "";
+        message.thinkingExpanded = true;
+      }
+      message.thinkingContent += event.token ?? "";
+      message.streaming = true;
+      scrollToBottom();
+    }
+  }
+
+  if (event.type === "done") {
+    const targetId = event.message_id || streamState.assistantMessageId;
+    if (!targetId) {
+      return;
+    }
+    const message = messages.value.find((item) => item.id === targetId);
+    if (message) {
+      message.streaming = false;
+    }
+  }
+
+  if (event.type === "tool_call_start" && event.call_id && event.name) {
+    const targetId = event.message_id || streamState.assistantMessageId;
+    if (!targetId) {
+      return;
+    }
+    const message = messages.value.find((item) => item.id === targetId);
+    if (message) {
+      if (!message.liveToolCalls) {
+        message.liveToolCalls = [];
+      }
+      message.liveToolCalls.push({
+        call_id: event.call_id,
+        name: event.name,
+        arguments: event.arguments,
+        done: false,
+      });
+      scrollToBottom();
+    }
+  }
+
+  if (event.type === "tool_call_result" && event.call_id) {
+    const targetId = event.message_id || streamState.assistantMessageId;
+    if (!targetId) {
+      return;
+    }
+    const message = messages.value.find((item) => item.id === targetId);
+    if (message?.liveToolCalls) {
+      const liveCall = message.liveToolCalls.find((item) => item.call_id === event.call_id);
+      if (liveCall) {
+        liveCall.result = event.result ?? "";
+        liveCall.done = true;
+        scrollToBottom();
+      }
+    }
+  }
+}
+
+async function sendMessage() {
+  await sendMessageWithText(draftMessage.value, false);
+}
+
+async function submitAskUserAnswer() {
+  await sendMessageWithText(askUserAnswer.value, true);
+}
+
+async function sendMessageWithText(rawInput: string, fromAskUser: boolean) {
+  if (sending.value) {
+    return;
+  }
+
+  const userText = rawInput.trim();
+  if (!userText) {
+    return;
+  }
+  if (!selectedService.value || selectedService.value.runtime.status !== "running") {
+    return;
+  }
+  if (!fromAskUser && !canSend.value) {
+    return;
+  }
+  if (fromAskUser && !canSubmitAskUser.value) {
+    return;
+  }
+  clearChatError();
+  const pendingNewConversation = parseNewConversationCommand(userText);
+  const requestMessages = [
+    ...toApiMessages(),
+    {
+      role: "user",
+      content: userText,
+    },
+  ];
+
+  if (fromAskUser) {
+    askUserAnswer.value = "";
+  } else {
+    draftMessage.value = "";
+  }
+  sending.value = true;
+
+  const streamState: StreamState = {
+    assistantMessageId: null,
+    pendingNewConversation,
+    requestText: userText,
+  };
+
+  if (!pendingNewConversation) {
+    const userMessage = {
+      id: `local-user-${crypto.randomUUID()}`,
+      role: "user" as const,
+      content: userText,
+      timestamp: new Date().toISOString(),
+      toolCalls: [],
+      toolCallId: null,
+      linkedToolCall: null,
+    };
+    messages.value.push(userMessage);
+
+    const assistantTempId = `local-assistant-${crypto.randomUUID()}`;
+    messages.value.push({
+      id: assistantTempId,
+      role: "assistant",
+      content: "",
+      streaming: true,
+      timestamp: new Date().toISOString(),
+      toolCalls: [],
+      toolCallId: null,
+      linkedToolCall: null,
+    });
+    streamState.assistantMessageId = assistantTempId;
+    scrollToBottom();
+  }
+
+  try {
+    await chat.stream(
+      {
+        agent_id: selectedServiceId.value,
+        session_id: activeSessionId.value || null,
+        stream: true,
+        model_config_id: selectedModelId.value || null,
+        thinking_type: selectedThinkingType.value || null,
+        reasoning_effort: selectedReasoningEffort.value || null,
+        workspace_path: isWorkspaceService.value ? workspacePath.value.trim() || null : null,
+        messages: requestMessages,
+      },
+      (event) => applyStreamEvent(event, streamState),
+    );
+    await reloadSessions();
+    if (activeSessionId.value) {
+      await openSession(activeSessionId.value);
+    }
+  } catch (error) {
+    applyInferenceFailure(streamState, (error as Error).message);
+  } finally {
+    sending.value = false;
+  }
+}
+
+async function load() {
+  servicesLoading.value = true;
+  try {
+    const [connections, llm, loadedAgents] = await Promise.all([
+      system.connections.list(),
+      system.llm.list(),
+      system.services.list(),
+    ]);
+    stats.connections = connections.length;
+    stats.llm = llm.length;
+    stats.services = loadedAgents.length;
+    services.value = loadedAgents;
+    llmModels.value = llm;
+
+    const eligible = loadedAgents.filter((agent) => chatEligibleServiceTypes.has(agent.agent_type.type));
+    const requestedAgent = props.agentId
+      ? loadedAgents.find((agent) => agent.config_id === props.agentId)
+      : null;
+
+    if (requestedAgent && chatEligibleServiceTypes.has(requestedAgent.agent_type.type)) {
+      selectedServiceId.value = requestedAgent.config_id;
+    } else if (
+      !selectedServiceId.value ||
+      !loadedAgents.some((agent) => agent.config_id === selectedServiceId.value)
+    ) {
+      selectedServiceId.value = eligible[0]?.config_id ?? loadedAgents[0]?.config_id ?? "";
+    }
+    selectedModelId.value = defaultAgentModelId.value;
+  } finally {
+    servicesLoading.value = false;
+  }
+
+  await reloadSessions();
+
+  if (props.sessionId) {
+    await openSession(props.sessionId).catch((error) => {
+      console.warn("Failed to open requested session:", error);
+    });
+  }
+}
+
+onMounted(() => {
+  load().catch((error) => {
+    console.error(error);
+    alert(`Chat 加载失败: ${(error as Error).message}`);
+  });
+  document.addEventListener("click", closePickersOnClickOutside);
+  document.addEventListener("keydown", handleToolPreviewKeydown);
+});
+
+onUnmounted(() => {
+  document.removeEventListener("click", closePickersOnClickOutside);
+  document.removeEventListener("keydown", handleToolPreviewKeydown);
+});
+</script>
+
+<style scoped lang="scss">
+.chat-page {
+  gap: 16px;
+  display: flex;
+  flex-direction: column;
+  height: calc(100vh - 40px);
+}
+
+.chat-page-panel,
+.chat-embedded-inner {
+  flex: 1;
+  min-height: 0;
+  display: flex;
+  flex-direction: column;
+}
+
+.chat-embedded-wrapper {
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+}
+
+.chat-embedded-wrapper .chat-panel {
+  flex: 1;
+}
+
+.chat-panel {
+  flex: 1;
+  min-height: 0;
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  padding: 24px;
+}
+
+.chat-toolbar {
+  display: flex;
+  gap: 16px;
+  justify-content: space-between;
+  align-items: flex-start;
+  padding-bottom: 16px;
+  border-bottom: 1px solid var(--admin-border);
+}
+
+.chat-agent-picker {
+  display: grid;
+  gap: 10px;
+  min-width: 420px;
+  flex: 1;
+}
+
+.chat-agent-picker-title {
+  margin-bottom: 0;
+  font-weight: 500;
+  color: var(--admin-muted);
+}
+
+.chat-agent-cards {
+  display: flex;
+  gap: 10px;
+  flex-wrap: wrap;
+}
+
+.chat-service-loading {
+  min-height: 54px;
+  display: flex;
+  align-items: center;
+  gap: 9px;
+  color: var(--admin-subtle);
+}
+
+.chat-service-loading-spinner {
+  width: 16px;
+  height: 16px;
+  border: 2px solid color-mix(in srgb, var(--admin-accent) 28%, transparent);
+  border-top-color: var(--admin-accent);
+  border-radius: 50%;
+  animation: spin 0.75s linear infinite;
+  flex-shrink: 0;
+}
+
+.chat-agent-card {
+  border: 1px solid var(--admin-border);
+  background: var(--admin-bg-soft);
+  border-radius: 12px;
+  padding: 8px 10px;
+  min-width: 220px;
+  display: flex;
+  align-items: center;
+  gap: 10px;
+  cursor: pointer;
+  transition: all 0.2s ease;
+}
+
+.chat-agent-card:hover {
+  border-color: color-mix(in srgb, var(--admin-accent) 32%, var(--admin-border) 68%);
+  transform: translateY(-1px);
+}
+
+.chat-agent-card.active {
+  border-color: color-mix(in srgb, var(--admin-accent) 52%, var(--admin-border) 48%);
+  box-shadow: 0 0 0 3px color-mix(in srgb, var(--admin-accent-soft) 70%, transparent 30%);
+  background: color-mix(in srgb, var(--admin-bg-elevated) 82%, var(--admin-accent-soft) 18%);
+}
+
+.chat-agent-card-avatar {
+  width: 40px;
+  height: 40px;
+  border-radius: 999px;
+  object-fit: cover;
+  border: 1px solid var(--admin-border);
+  flex-shrink: 0;
+}
+
+.chat-agent-card-avatar--fallback {
+  display: grid;
+  place-items: center;
+  background: color-mix(in srgb, var(--admin-accent) 18%, var(--admin-bg-panel) 82%);
+  color: var(--admin-ink);
+  font-weight: 700;
+}
+
+.chat-agent-card-meta {
+  display: grid;
+  text-align: left;
+  min-width: 0;
+}
+
+.chat-agent-card-meta strong {
+  font-size: 14px;
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+}
+
+.chat-agent-card-meta span {
+  font-size: 12px;
+  color: var(--admin-subtle);
+}
+
+.chat-agent-card.inactive {
+  opacity: 0.55;
+  border-color: var(--admin-border);
+}
+
+.chat-agent-card.inactive:hover {
+  opacity: 0.8;
+}
+
+.agent-status-badge {
+  background: var(--admin-danger, #ef4444);
+  color: white;
+  font-size: 11px;
+  font-weight: 600;
+  padding: 2px 7px;
+  border-radius: 5px;
+  margin-left: auto;
+  flex-shrink: 0;
+}
+
+.chat-layout {
+  margin-top: 16px;
+  display: grid;
+  grid-template-columns: 240px 1fr;
+  gap: 20px;
+  flex: 1;
+  min-height: 0;
+}
+
+.chat-sessions {
+  background: var(--admin-bg-panel);
+  border: 1px solid var(--admin-border);
+  border-radius: 16px;
+  padding: 14px;
+  min-height: 0;
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+  overflow-y: auto;
+}
+
+.chat-sessions-header {
+  font-size: 13px;
+  font-weight: 600;
+  color: var(--admin-muted);
+  padding: 4px 8px;
+  margin-bottom: 4px;
+  text-transform: uppercase;
+  letter-spacing: 0.5px;
+}
+
+.chat-session-item {
+  border: 1px solid transparent;
+  background: transparent;
+  color: var(--admin-ink);
+  border-radius: 10px;
+  padding: 6px 8px 6px 14px;
+  text-align: left;
+  display: flex;
+  align-items: center;
+  gap: 4px;
+  cursor: pointer;
+  transition: all 0.2s ease;
+}
+
+.chat-session-item:hover {
+  background: var(--admin-bg-soft);
+}
+
+.chat-session-item.active {
+  background: var(--admin-bg-elevated);
+  border-color: var(--admin-border);
+  box-shadow: 0 4px 12px color-mix(in srgb, var(--admin-bg) 60%, transparent);
+}
+
+.chat-session-main {
+  flex: 1;
+  min-width: 0;
+  display: flex;
+  flex-direction: column;
+  gap: 6px;
+  background: transparent;
+  border: none;
+  padding: 0;
+  color: inherit;
+  text-align: left;
+  cursor: pointer;
+}
+
+.chat-session-main strong {
+  font-size: 14px;
+  font-family: monospace;
+}
+
+.chat-session-main .muted {
+  font-size: 12px;
+  color: var(--admin-subtle);
+}
+
+.chat-session-delete {
+  width: 24px;
+  height: 24px;
+  border-radius: 6px;
+  border: none;
+  background: transparent;
+  color: var(--admin-subtle);
+  font-size: 16px;
+  line-height: 1;
+  cursor: pointer;
+  display: grid;
+  place-items: center;
+  flex-shrink: 0;
+  opacity: 0;
+  transition: all 0.2s ease;
+}
+
+.chat-session-item:hover .chat-session-delete {
+  opacity: 1;
+}
+
+.chat-session-delete:hover {
+  background: color-mix(in srgb, var(--admin-danger, #ef4444) 12%, transparent);
+  color: var(--admin-danger, #ef4444);
+}
+
+.chat-main {
+  background: var(--admin-bg-panel);
+  border: 1px solid var(--admin-border);
+  border-radius: 16px;
+  min-height: 0;
+  display: flex;
+  flex-direction: column;
+  overflow: hidden;
+  box-shadow: 0 4px 24px color-mix(in srgb, var(--admin-bg) 60%, transparent);
+}
+
+.chat-messages {
+  flex: 1;
+  min-height: 0;
+  padding: 24px;
+  overflow-y: auto;
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+}
+
+.chat-bubble-row {
+  display: flex;
+  align-items: flex-end;
+  gap: 10px;
+}
+
+.chat-bubble-row.user {
+  justify-content: flex-end;
+}
+
+.chat-bubble-row.assistant {
+  justify-content: flex-start;
+}
+
+.chat-bubble-col {
+  display: flex;
+  flex-direction: column;
+  max-width: calc(85% - 54px);
+}
+
+.chat-tool-above-content {
+  margin-bottom: 6px;
+}
+
+.chat-tool-below-content {
+  margin-top: 6px;
+}
+
+.chat-message-item + .chat-message-item {
+  margin-top: 8px;
+}
+
+.chat-message-avatar {
+  width: 40px;
+  height: 40px;
+  border-radius: 999px;
+  object-fit: cover;
+  border: 1px solid var(--admin-border);
+  flex-shrink: 0;
+  margin-bottom: 2px;
+}
+
+.chat-message-avatar--fallback {
+  display: grid;
+  place-items: center;
+  background: color-mix(in srgb, var(--admin-accent) 16%, var(--admin-bg-panel) 84%);
+  color: var(--admin-ink);
+  font-size: 13px;
+  font-weight: 700;
+}
+
+.chat-bubble {
+  max-width: 85%;
+  border-radius: 18px;
+  padding: 12px 18px;
+  line-height: 1.6;
+  font-size: 15px;
+  box-shadow: 0 2px 8px color-mix(in srgb, var(--admin-bg) 40%, transparent);
+  overflow-wrap: anywhere;
+}
+
+.chat-bubble-col .chat-bubble {
+  max-width: 100%;
+}
+
+.chat-thinking-block {
+  margin-bottom: 10px;
+  border: 1px solid var(--admin-border);
+  border-radius: 10px;
+  overflow: hidden;
+  background: color-mix(in srgb, var(--admin-accent) 6%, var(--admin-bg-panel) 94%);
+}
+
+.chat-thinking-block.collapsed {
+  background: transparent;
+  border-color: transparent;
+}
+
+.chat-thinking-toggle {
+  display: flex;
+  align-items: center;
+  gap: 6px;
+  width: 100%;
+  background: none;
+  border: none;
+  padding: 6px 10px;
+  font-size: 13px;
+  color: var(--admin-subtle);
+  cursor: pointer;
+  text-align: left;
+  line-height: 1.4;
+}
+
+.chat-thinking-toggle:hover {
+  color: var(--admin-accent);
+}
+
+.chat-thinking-icon {
+  font-size: 10px;
+  flex-shrink: 0;
+}
+
+.chat-thinking-content {
+  padding: 8px 12px 10px;
+  font-size: 13px;
+  color: var(--admin-subtle);
+  line-height: 1.5;
+  white-space: pre-wrap;
+  border-top: 1px solid var(--admin-border);
+}
+
+.chat-bubble-time {
+  margin-top: 6px;
+  font-size: 12px;
+  color: var(--admin-subtle);
+  opacity: 0.85;
+  text-align: right;
+}
+
+.chat-bubble.assistant .chat-bubble-time {
+  text-align: left;
+}
+
+.chat-bubble.user {
+  background: var(--admin-accent);
+  color: #fff;
+  border-bottom-right-radius: 4px;
+}
+
+.chat-bubble.assistant {
+  background: var(--admin-bg-elevated);
+  border: 1px solid var(--admin-border);
+  color: var(--admin-ink);
+  border-bottom-left-radius: 4px;
+}
+
+.chat-tool-inline-list {
+  margin-top: 8px;
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  gap: 4px;
+}
+
+.chat-tool-inline {
+  border: none;
+  background: transparent;
+  padding: 0;
+  font-size: 12px;
+  line-height: 1.4;
+  color: var(--admin-subtle);
+  cursor: pointer;
+  text-align: left;
+  display: flex;
+  align-items: center;
+  gap: 5px;
+}
+
+.chat-tool-inline:hover,
+.chat-tool-inline.active {
+  color: var(--admin-accent);
+  text-decoration: underline;
+}
+
+.tool-preview-overlay {
+  position: fixed;
+  inset: 0;
+  background: rgba(0, 0, 0, 0.55);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  z-index: 10000;
+  backdrop-filter: blur(2px);
+}
+
+.tool-preview-panel {
+  background: var(--admin-bg-elevated);
+  border: 1px solid var(--admin-border);
+  border-radius: 12px;
+  width: 90%;
+  max-width: 800px;
+  max-height: 80vh;
+  display: flex;
+  flex-direction: column;
+  box-shadow: 0 12px 40px rgba(0, 0, 0, 0.4);
+  overflow: hidden;
+}
+
+.tool-preview-header {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+  padding: 14px 18px;
+  border-bottom: 1px solid var(--admin-border);
+  font-size: 14px;
+  font-weight: 600;
+  color: var(--admin-ink);
+  flex-shrink: 0;
+}
+
+.tool-preview-close {
+  margin-left: auto;
+  border: none;
+  background: transparent;
+  color: var(--admin-subtle);
+  font-size: 16px;
+  cursor: pointer;
+  padding: 2px 6px;
+  border-radius: 4px;
+  line-height: 1;
+}
+
+.tool-preview-close:hover {
+  color: var(--admin-accent);
+  background: color-mix(in srgb, var(--admin-accent-soft) 30%, transparent);
+}
+
+.tool-preview-body {
+  padding: 16px 18px;
+  overflow-y: auto;
+  flex: 1;
+  min-height: 0;
+}
+
+.tool-preview-code {
+  margin: 0;
+  padding: 14px 16px;
+  border-radius: 8px;
+  font-size: 13px;
+  line-height: 1.5;
+  white-space: pre-wrap;
+  word-break: break-word;
+  font-family: "Cascadia Code", "Fira Code", "JetBrains Mono", monospace;
+  max-height: 60vh;
+  overflow: auto;
+}
+
+.tool-preview-code--create {
+  background: color-mix(in srgb, var(--admin-good) 10%, var(--admin-bg) 90%);
+  border: 1px solid color-mix(in srgb, var(--admin-good) 25%, transparent);
+}
+
+.tool-preview-code--cmd {
+  background: color-mix(in srgb, var(--admin-bg-panel-strong) 70%, var(--admin-ink) 30%);
+  border: 1px solid var(--admin-border-strong);
+  color: color-mix(in srgb, var(--admin-ink) 90%, var(--admin-good) 10%);
+}
+
+.tool-preview-code--error {
+  color: color-mix(in srgb, var(--admin-bad) 85%, var(--admin-ink) 15%);
+}
+
+.tool-preview-info {
+  padding: 20px;
+  border-radius: 8px;
+  font-size: 14px;
+  line-height: 1.6;
+}
+
+.tool-preview-info--delete {
+  background: color-mix(in srgb, var(--admin-bad) 10%, var(--admin-bg) 90%);
+  border: 1px solid color-mix(in srgb, var(--admin-bad) 25%, transparent);
+}
+
+.tool-preview-info p {
+  margin: 0 0 8px;
+}
+
+.tool-preview-info p:last-child {
+  margin-bottom: 0;
+}
+
+.tool-preview-info code {
+  background: color-mix(in srgb, var(--admin-bad) 15%, var(--admin-bg) 85%);
+  padding: 2px 6px;
+  border-radius: 4px;
+  font-family: "Cascadia Code", "Fira Code", "JetBrains Mono", monospace;
+  font-size: 13px;
+}
+
+.tool-preview-diff {
+  display: flex;
+  flex-direction: column;
+  gap: 12px;
+}
+
+.tool-preview-hunk {
+  border-radius: 8px;
+  overflow: hidden;
+  border: 1px solid var(--admin-border);
+}
+
+.tool-preview-diff-line {
+  padding: 2px 12px 2px 8px;
+  font-size: 13px;
+  line-height: 1.5;
+  font-family: "Cascadia Code", "Fira Code", "JetBrains Mono", monospace;
+  white-space: pre-wrap;
+  word-break: break-all;
+}
+
+.tool-preview-diff-line .diff-marker {
+  display: inline-block;
+  width: 16px;
+  text-align: center;
+  font-weight: 700;
+  user-select: none;
+  margin-right: 4px;
+}
+
+.tool-preview-diff-line--removed {
+  background: color-mix(in srgb, var(--admin-bad) 14%, var(--admin-bg) 86%);
+  color: var(--admin-bad);
+}
+
+.tool-preview-diff-line--added {
+  background: color-mix(in srgb, var(--admin-good) 14%, var(--admin-bg) 86%);
+  color: var(--admin-good);
+}
+
+.tool-preview-output {
+  margin-bottom: 14px;
+}
+
+.tool-preview-output:last-child {
+  margin-bottom: 0;
+}
+
+.tool-preview-output-label {
+  font-size: 11px;
+  font-weight: 600;
+  text-transform: uppercase;
+  letter-spacing: 0.5px;
+  color: var(--admin-subtle);
+  margin-bottom: 6px;
+}
+
+.tool-preview-output-label--error {
+  color: var(--admin-bad);
+}
+
+.tool-preview-no-result {
+  font-size: 14px;
+  color: var(--admin-subtle);
+  font-style: italic;
+  text-align: center;
+  padding: 24px;
+}
+
+.chat-live-tool-wrapper {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  width: 100%;
+}
+
+.live-tool-spinner {
+  display: inline-block;
+  width: 10px;
+  height: 10px;
+  border: 2px solid color-mix(in srgb, var(--admin-accent) 40%, transparent);
+  border-top-color: var(--admin-accent);
+  border-radius: 50%;
+  animation: spin 0.7s linear infinite;
+  flex-shrink: 0;
+}
+
+.live-tool-done-icon {
+  color: var(--admin-accent);
+  font-size: 11px;
+  font-weight: 700;
+  flex-shrink: 0;
+}
+
+.live-tool-pending {
+  font-size: 12px;
+  color: var(--admin-subtle);
+  font-style: italic;
+}
+
+@keyframes spin {
+  to {
+    transform: rotate(360deg);
+  }
+}
+
+.chat-tool-detail-inline {
+  margin-top: 10px;
+  padding: 12px 14px;
+  background: color-mix(in srgb, var(--admin-bg-panel) 74%, var(--admin-accent-soft) 26%);
+  border: 1px dashed color-mix(in srgb, var(--admin-accent) 42%, var(--admin-border) 58%);
+  border-radius: 12px;
+  display: grid;
+  gap: 10px;
+}
+
+.chat-tool-detail-inline-header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 12px;
+}
+
+.chat-tool-detail-inline-close {
+  border: none;
+  background: transparent;
+  padding: 0;
+  font-size: 12px;
+  color: var(--admin-subtle);
+  cursor: pointer;
+}
+
+.chat-tool-detail-inline-close:hover {
+  color: var(--admin-accent);
+}
+
+.chat-tool-detail-caption {
+  font-size: 12px;
+  color: var(--admin-subtle);
+  margin-bottom: 4px;
+}
+
+.chat-tool-detail-inline-block {
+  min-width: 0;
+}
+
+.chat-tool-detail-inline pre {
+  margin: 0;
+  white-space: pre-wrap;
+  word-break: break-word;
+  background: color-mix(in srgb, var(--admin-bg) 82%, black 18%);
+  border: 1px solid var(--admin-border);
+  border-radius: 12px;
+  padding: 12px 14px;
+  max-height: 180px;
+  overflow: auto;
+}
+
+.chat-input-area {
+  border-top: 1px solid var(--admin-border);
+  padding: 16px 20px;
+  display: flex;
+  flex-direction: column;
+  gap: 12px;
+  background: var(--admin-bg-elevated);
+}
+
+.ask-user-panel {
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+  padding: 12px 14px;
+  border: 1px solid var(--admin-border);
+  border-radius: 12px;
+  background: color-mix(in srgb, var(--admin-bg-panel) 88%, var(--admin-bg) 12%);
+}
+
+.ask-user-question {
+  font-size: 13px;
+  font-weight: 700;
+  color: var(--admin-ink);
+}
+
+.ask-user-row input {
+  background: var(--admin-bg);
+  color: var(--admin-ink);
+  border: 1px solid var(--admin-border);
+  border-radius: 10px;
+  padding: 10px 12px;
+}
+
+.workspace-path-display {
+  padding: 10px 16px;
+  background: var(--admin-bg-elevated);
+  border-bottom: 1px solid var(--admin-border);
+  display: flex;
+  align-items: center;
+  gap: 8px;
+}
+
+.workspace-path-display .path-label {
+  font-size: 13px;
+  color: var(--admin-subtle);
+  font-weight: 500;
+  white-space: nowrap;
+}
+
+.workspace-path-display .path-value {
+  font-size: 13px;
+  color: var(--admin-ink);
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+
+.workspace-path-display .path-value.path-unset {
+  color: var(--admin-muted);
+  font-style: italic;
+}
+
+.chat-session-group-header {
+  font-size: 12px;
+  color: var(--admin-muted);
+  padding: 6px 10px 4px;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+  border-bottom: 1px solid var(--admin-border);
+  margin-top: 4px;
+}
+
+.chat-session-group-header:first-child {
+  margin-top: 0;
+}
+
+.ask-user-details {
+  color: var(--admin-muted);
+  font-size: 13px;
+  line-height: 1.5;
+  white-space: pre-wrap;
+}
+
+.ask-user-row {
+  display: flex;
+  gap: 10px;
+  flex-wrap: wrap;
+}
+
+.ask-user-row input {
+  flex: 1 1 320px;
+}
+
+.chat-input-area textarea {
+  background: var(--admin-bg);
+  color: var(--admin-ink);
+  border: 1px solid var(--admin-border);
+  resize: none;
+  height: 80px;
+  border-radius: 12px;
+  padding: 12px 16px;
+  font-size: 15px;
+  line-height: 1.5;
+  transition: all 0.2s;
+}
+
+.chat-input-area textarea:focus {
+  border-color: var(--admin-accent);
+  outline: none;
+  box-shadow: 0 0 0 3px var(--admin-accent-soft);
+}
+
+.chat-not-supported {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  gap: 8px;
+  padding: 32px 20px;
+  text-align: center;
+  color: var(--admin-muted);
+}
+
+.chat-not-supported-icon {
+  font-size: 32px;
+  line-height: 1;
+  opacity: 0.7;
+}
+
+.chat-not-supported-title {
+  font-size: 15px;
+  font-weight: 600;
+  color: var(--admin-ink);
+}
+
+.chat-not-supported-desc {
+  font-size: 13px;
+  color: var(--admin-subtle);
+}
+
+.chat-input-actions {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  gap: 10px;
+}
+
+.chat-error-box {
+  padding: 12px 14px;
+  border-radius: 12px;
+  border: 1px solid color-mix(in srgb, var(--admin-danger, #ef4444) 40%, var(--admin-border) 60%);
+  background: color-mix(in srgb, var(--admin-danger, #ef4444) 12%, var(--admin-bg-panel) 88%);
+  color: color-mix(in srgb, var(--admin-danger, #ef4444) 82%, var(--admin-ink) 18%);
+  font-size: 13px;
+  line-height: 1.5;
+  white-space: pre-wrap;
+}
+
+.chat-input-right {
+  display: flex;
+  align-items: center;
+  gap: 10px;
+  flex-wrap: wrap;
+}
+
+.chat-model-bar {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+}
+
+.model-picker {
+  position: relative;
+}
+
+.model-chip {
+  display: inline-flex;
+  align-items: center;
+  gap: 6px;
+  background: var(--admin-bg);
+  border: 1px solid var(--admin-border);
+  border-radius: 10px;
+  padding: 6px 12px;
+  font-size: 13px;
+  font-weight: 500;
+  color: var(--admin-ink);
+  cursor: pointer;
+  transition: all 0.2s ease;
+  white-space: nowrap;
+  line-height: 1;
+}
+
+.model-chip:hover {
+  border-color: color-mix(in srgb, var(--admin-accent) 40%, var(--admin-border) 60%);
+  background: var(--admin-bg-elevated);
+}
+
+.model-chip.icon-only {
+  padding: 6px;
+  color: var(--admin-muted);
+}
+
+.model-chip.icon-only:hover {
+  color: var(--admin-accent);
+}
+
+.model-chip svg {
+  width: 14px;
+  height: 14px;
+  flex-shrink: 0;
+}
+
+.chip-chevron {
+  transition: transform 0.2s ease;
+}
+
+.model-picker.open .chip-chevron {
+  transform: rotate(180deg);
+}
+
+.model-picker-dropdown {
+  position: absolute;
+  bottom: calc(100% + 6px);
+  left: 0;
+  min-width: 200px;
+  max-height: 280px;
+  overflow-y: auto;
+  background: var(--admin-bg-panel);
+  border: 1px solid var(--admin-border);
+  border-radius: 12px;
+  padding: 6px;
+  display: flex;
+  flex-direction: column;
+  gap: 2px;
+  box-shadow: 0 8px 24px color-mix(in srgb, var(--admin-bg) 60%, transparent);
+  z-index: 10;
+}
+
+.model-picker-item {
+  background: transparent;
+  border: none;
+  border-radius: 8px;
+  padding: 8px 10px;
+  font-size: 13px;
+  color: var(--admin-ink);
+  text-align: left;
+  cursor: pointer;
+  transition: all 0.15s ease;
+  white-space: nowrap;
+}
+
+.model-picker-item:hover {
+  background: var(--admin-bg-soft);
+}
+
+.model-picker-item.active {
+  background: color-mix(in srgb, var(--admin-accent) 12%, var(--admin-bg-soft) 88%);
+  color: var(--admin-accent);
+  font-weight: 600;
+}
+
+.model-settings {
+  position: relative;
+}
+
+.chat-bubble-content :deep(*) {
+  margin: 0;
+}
+
+.chat-bubble-content :deep(* + *) {
+  margin-top: 0.75em;
+}
+
+.chat-bubble-content :deep(p),
+.chat-bubble-content :deep(li) {
+  white-space: pre-wrap;
+}
+
+.chat-bubble-content :deep(ul),
+.chat-bubble-content :deep(ol) {
+  padding-left: 1.4em;
+}
+
+.chat-bubble-content :deep(blockquote) {
+  padding-left: 12px;
+  border-left: 3px solid color-mix(in srgb, var(--admin-accent) 45%, transparent);
+  color: var(--admin-subtle);
+}
+
+.chat-bubble-content :deep(code) {
+  font-family: "Cascadia Code", "JetBrains Mono", Consolas, monospace;
+  font-size: 0.92em;
+  padding: 0.15em 0.4em;
+  border-radius: 6px;
+  background: color-mix(in srgb, var(--admin-bg) 82%, black 18%);
+}
+
+.chat-bubble-content :deep(pre) {
+  overflow-x: auto;
+  padding: 12px 14px;
+  border-radius: 12px;
+  border: 1px solid var(--admin-border);
+  background: color-mix(in srgb, var(--admin-bg) 82%, black 18%);
+}
+
+.chat-bubble-content :deep(pre code) {
+  display: block;
+  padding: 0;
+  background: transparent;
+  white-space: pre;
+}
+
+.chat-bubble-content :deep(a) {
+  color: inherit;
+  text-decoration: underline;
+}
+
+.chat-bubble-content :deep(hr) {
+  border: 0;
+  border-top: 1px solid var(--admin-border);
+}
+
+@media (max-width: 1180px) {
+  .chat-toolbar {
+    flex-direction: column;
+    align-items: stretch;
+  }
+
+  .chat-agent-picker {
+    min-width: 0;
+  }
+
+  .chat-layout {
+    grid-template-columns: 1fr;
+  }
+
+  .chat-sessions {
+    max-height: 240px;
+  }
+}
+
+@media (max-width: 900px) {
+  .chat-page {
+    height: auto;
+    min-height: calc(100vh - 32px);
+  }
+
+  .chat-panel {
+    padding: 18px;
+  }
+
+  .chat-agent-card {
+    min-width: min(220px, 100%);
+    flex: 1 1 220px;
+  }
+
+  .chat-messages {
+    padding: 18px;
+  }
+
+  .chat-input-area {
+    padding: 14px 16px;
+  }
+
+  .chat-input-actions {
+    flex-wrap: wrap;
+    justify-content: flex-end;
+  }
+
+  .chat-bubble,
+  .chat-bubble-col .chat-bubble {
+    max-width: 100%;
+  }
+}
+</style>
